Cin 61 back 리뷰작성 api

+++ src/main/java/com/back/cinetalk/movie/dto/MovieDTO.java
@@ -18,7 +18,7 @@ public class MovieDTO {
 
     private Long id;
 
-    private int movieId;
+    private Long movieId;
 
     private String movienm;
